% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_helper.R
\name{batch_mutate}
\alias{batch_mutate}
\title{Batch Mutate}
\usage{
batch_mutate(df, df_col, system_prompt, batch_size = 100)
}
\arguments{
\item{df}{A data frame.}

\item{df_col}{A column from the data frame.}

\item{system_prompt}{A system prompt for the GPT model.}

\item{batch_size}{The number of rows to process in each batch. Default is 100.}
}
\value{
The final output data frame after processing all the rows.
}
\description{
This function applies the \code{mutate_row} function in batches to a data frame.
It processes a specified number of rows at a time and saves the progress.
It returns the final output data frame after all the rows have been processed.
}
