% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_batch.R
\name{gpt_batch}
\alias{gpt_batch}
\title{GPT Batch}
\usage{
gpt_batch(
  df,
  input,
  prompt,
  batch_size = 10,
  attempts = 1,
  model = "gpt-3.5-turbo-1106",
  temperature = 0.1
)
}
\arguments{
\item{df}{A data frame (to preserve all columns).}

\item{input}{An input vector from \code{df} (to generate the completions).}

\item{prompt}{A system prompt for the GPT model.}

\item{batch_size}{The number of rows to process in each batch. Default is 10.}

\item{attempts}{The maximum number of attempts in case of errors. Default is 1.}

\item{model}{An OpenAI model. Default is "gpt-3.5-turbo-1106".}

\item{temperature}{A temperature for the GPT model. Default is 0.1.}
}
\value{
Writes the GPT completion to \code{gpt_output.RDS}. Writes the progress to \code{last_completed_batch.RDS}.
}
\description{
Using OpenAI's GPT models, generate a completion in a new data frame column called \code{gpt_output}.
Completions are batched for handling errors and saving your progress.
For API etiquette, there is a delay between each completion and each batch.
}
\examples{
library(dylanR)

Sys.setenv(OPENAI_API_KEY = "...")

objects <- dplyr::tibble(user = c("chair", "table", "book", "pen", "lamp", "phone", "keyboard", "door", "window", "cup"))

gpt_batch(
  objects,
  objects$user,
  "write a one sentence description"
)

objects_described <- readRDS("gpt_output.RDS")
}
