% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_helper.R
\name{gpt_helper}
\alias{gpt_helper}
\title{GPT Helper}
\usage{
gpt_helper(
  df,
  input,
  prompt,
  batch_size = 10,
  retries = 3,
  model = "gpt-3.5-turbo",
  temperature = 0.1
)
}
\arguments{
\item{df}{A data frame (to preserve all columns).}

\item{input}{An input vector from \code{df} (to generate the completion).}

\item{prompt}{A system prompt for the GPT model.}

\item{batch_size}{The number of rows to process in each batch. Default is 10.}

\item{retries}{The maximum number of retries in case of errors. Default is 3.}

\item{model}{A GPT model. Default is "gpt-3.5-turbo".}

\item{temperature}{A temperature for the GPT model. Default is 0.1.}
}
\value{
Writes the GPT completion to \code{gpt_output.RDS}.
}
\description{
This function uses OpenAI's GPT-3.5-Turbo model to generate a completion inside of a data frame.
}
\examples{
library(dylanR)

Sys.setenv(OPENAI_API_KEY = "...")

objects <- dplyr::tibble(user = c("chair", "table", "book", "pen", "lamp", "phone", "keyboard", "door", "window", "cup"))

gpt_helper(
  objects,
  objects$user,
  "write a one sentence description"
)

objects_described <- readRDS("gpt_output.Rds")
}
